Moved getObjectIDTargetTypeResolvers to OverridingTrait
